still working on room DB

+++ app/src/main/java/core/Interfaces.kt
@@ -1,5 +1,8 @@
 package core
 
+import data.Card
+import data.Member
+
 interface ICardService {
     fun findCardById(cardId: String): Card?
     fun findCardByIdAndType(cardId: String, cardType: String): Card?
@@ -7,5 +10,5 @@ interface ICardService {
 
 interface IMemberService{
     fun getMembers(): MutableList<Member>
-    fun getMemberById(memberId: Int): Member?
+    fun getMemberById(memberId: Long): Member?
 }
