+++ app/src/main/java/data/MemberRepository.kt
@@ -1,13 +1,11 @@
 package data
 
-import core.Member
-
 class MemberRepository : IMemberRepository {
     override fun selectMembers(): MutableList<Member> {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
-    override fun selectMemberById(memberId: Int): Member? {
+    override fun selectMemberById(memberId: Long): Member? {
         var foundMember: Member? = null
         InMemoryData.members.forEach{
             if(it.id == memberId) {
